+++ sentimental_analysis/realworld/Scraper/etsy_scrapper.py
