+++ sentimental_analysis/realworld/Scraper/tripadvisor_scrapper.py
